+++ Sources/LCP/Content Protection/EncryptionParser.swift
@@ -51,11 +51,7 @@ private func parseEPUBEncryptionData(in container: Container) async -> ReadResul
             do {
                 let doc = try await DefaultXMLDocumentFactory().open(
                     data: data,
-                    namespaces: [
-                        (prefix: "enc", uri: "http://www.w3.org/2001/04/xmlenc#"),
-                        (prefix: "ds", uri: "http://www.w3.org/2000/09/xmldsig#"),
-                        (prefix: "comp", uri: "http://www.idpf.org/2016/encryption#compression"),
-                    ]
+                    namespaces: [.enc, .ds, .comp]
                 )
                 return .success(doc)
             } catch {