+++ CHANGELOG.md
@@ -12,6 +12,10 @@ All notable changes to this project will be documented in this file. Take a look
 
 * Implementation of the [W3C Accessibility Metadata Display Guide](https://w3c.github.io/publ-a11y/a11y-meta-display-guide/2.0/guidelines/) specification to facilitate displaying accessibility metadata to users. [See the dedicated user guide](docs/Guides/Accessibility.md).
 * Support for starting from a progression in the HTML content iterator.
+* New link `rels` in the `readingOrder` and EPUB `landmarks` to mark:
+    * `cover`: the title page
+    * `contents`: the table of contents
+    * `start`: the first actual chapter
 
 #### Navigator